# api/dumprun_openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dump Pickup API",
    "version": "1.0.0",
    "description": "API for managing dump pickup requests, user accounts, and driver operations."
  },
  "servers": [
    {
      "url": "https://api.dumppickup.com/v1"
    }
  ],
  "paths": {
...
    "/v1/pickups": {
      "post": {
        "summary": "Create a new pickup request",
        "operationId": "createPickup",
        "tags": ["Pickups"],
        "security": [
          {
            "cognitoAuth": ["openid"]
          }
        ],
        "x-amazon-apigateway-auth": {
          "type": "cognito_user_pools"
        },
        "x-amazon-apigateway-integration": {
          "type": "aws_proxy",
          "uri": "arn:aws:apigateway:region:lambda:path/2015-03-31/functions/arn:aws:lambda:region:account-id:function:create-pickup/invocations",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "POST"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPickup"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pickup request created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pickup"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List pickup requests",
        "operationId": "listPickups",
        "tags": ["Pickups"],
        "security": [
          {
            "CognitoAuth": ["admin"]
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["pending", "assigned", "completed", "cancelled"]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pickup requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pickups": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pickup"
                      }
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/pickups/{pickupId}": {
      "get": {
        "summary": "Get a pickup by ID",
        "operationId": "getPickup",
        "tags": ["Pickups"],
        "security": [
          {
            "CognitoAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pickupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pickup details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pickup"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Update a pickup",
        "operationId": "updatePickup",
        "tags": ["Pickups"],
        "security": [
          {
            "CognitoAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pickupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePickup"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pickup updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pickup"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete a pickup",
        "operationId": "deletePickup",
        "tags": ["Pickups"],
        "security": [
          {
            "CognitoAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pickupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pickup deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/v1/pickups/{pickupId}/accept": {
      "post": {
        "summary": "Accept a pickup request",
        "operationId": "acceptPickup",
        "tags": ["Pickups"],
        "security": [
          {
            "CognitoAuth": ["driver"]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pickupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pickup request accepted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pickup"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel acceptance of a pickup",
        "operationId": "cancelAcceptance",
        "tags": ["Pickups"],
        "security": [
          {
            "CognitoAuth": ["driver"]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pickupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Acceptance cancelled successfully"
          },
          "404": {
            "description": "Pickup not found or not currently accepted"
          },
          "403": {
            "description": "Not authorized to cancel this acceptance"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "CognitoAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "x-amazon-apigateway-authtype": "cognito_user_pools",
        "x-amazon-apigateway-authorizer": {
          "type": "cognito_user_pools",
          "providerARNs": [
            "arn:aws:cognito-idp:us-west-2:369547265320:userpool/us-west-2_vriluhdDN"
          ]
        }
      }
    },
    "schemas": {
...
      "NewPickup": {
        "type": "object",
        "required": [
          "location",
          "estimatedWeight",
          "requestedTime",
          "wasteType"
        ],
        "properties": {
          "location": {
            "type": "string"
          },
          "estimatedWeight": {
            "type": "number",
            "format": "float"
          },
          "wasteType": {
            "type": "string"
          },
          "requestedTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdatePickup": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string"
          },
          "estimatedWeight": {
            "type": "number",
            "format": "float",
            "minimum": 1
          },
          "wasteType": {
            "type": "string",
            "enum": ["household", "construction", "green", "electronic"]
          },
          "requestedTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "available",
              "accepted",
              "in_progress",
              "completed",
              "cancelled"
            ]
          }
        }
      },
      "Pickup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "driverId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "available",
              "accepted",
              "in_progress",
              "completed",
              "cancelled"
            ]
          },
          "location": {
            "type": "string"
          },
          "estimatedWeight": {
            "type": "number",
            "format": "float",
            "minimum": 1
          },
          "wasteType": {
            "type": "string",
            "enum": ["household", "construction", "green", "electronic"]
          },
          "requestedTime": {
            "type": "string",
            "format": "date-time"
          },
          "assignedTime": {
            "type": "string",
            "format": "date-time"
          },
          "completedTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}